+++ wrapper.h
@@ -10,6 +10,6 @@
 
 void forward_wrapper(double * input, double * bias, int output_dim, int input_dim, double * intermediate, double * weights, double * output, Activation_Function *activation_function);
 
-void backward_wrapper(double * actual_outputs, double * bias, Cost_Function * f, int learning_rate, int input_dim, int output_dim, double * intermediate, double * weights, double * output, Activation_Function *activation_function, bool final_layer, double * error_term);
+void backward_wrapper(double * actual_outputs, double * bias, Cost_Function * f, double learning_rate, int input_dim, int output_dim, double * intermediate, double * weights, double * output, Activation_Function *activation_function, bool final_layer, double * error_term);
 
 #endif
